pyproject.toml
[project]
name = "microwave-ai"
version = "0.3.0"
description = "Decentralized AI inference with pipeline parallelism, speculative decoding, and latency-optimized routing."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "websockets>=12.0",
  "numpy>=1.24.0",
  "psutil>=5.9.0",
  "msgpack>=1.0.0",
  "lz4>=4.3.0",
]

[project.optional-dependencies]
pipeline = [
  "llama-cpp-python>=0.2.0",
]

[project.scripts]
microwave-gateway = "microwave_ai.gateway:main"
microwave-node = "microwave_ai.node:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["microwave_ai*"]
